+++ CHANGELOG.md
@@ -118,6 +118,7 @@ Plus support for **CogView-4**, **SANA 1.5**, new CLiP models, improvements to r
   - fix legacy diffusion latent upscalers  
   - fix upscaler selection in postprocessing  
   - fix sd35 with batch processing  
+  - fix extra networks cover and inline views  
 
 ## Update for 2025-02-28